server/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
python-dotenv==1.1.0
Werkzeug==3.1.3
pytest==8.3.5
pytest-flask==1.3.0
